refactor: directly access, no using props
