+++ components/main/index.js
@@ -10,27 +10,13 @@ import { Categories } from "./Categories.js";
  * @async
  * @function Main
  *
- * @param {Promise<RankItem[]>} top10SeriesPromise
- * @param {Promise<Item[]>} newContentsPromise
- * @param {Promise<RankItem[]>} top10MoviesPromise
- *
  * @returns {Promise<string>}
  */
-export async function Main(
-  top10SeriesPromise,
-  newContentsPromise,
-  top10MoviesPromise
-) {
-  const categories = await Categories(
-    top10SeriesPromise,
-    newContentsPromise,
-    top10MoviesPromise
-  );
-
+export async function Main() {
   return `
 <main>
   ${Hero()}
-  ${categories}
+  ${await Categories()}
 </main>
 `.trim();
 }
